started refactoring to reuse list and cards
