+++ projectAtelier/client/src/components/RelatedProducts/RelatedSubs/Comparison.jsx
@@ -6,7 +6,6 @@ import TableBody from '@mui/material/TableBody';
 import convertPrice from '../../../Middleware/convertPrice.js';
 import CheckIcon from '@mui/icons-material/Check';
 import CloseIcon from '@mui/icons-material/Close';
-
 import '../../../index.css';
