--- projectAtelier/client/src/components/RelatedProducts/RelatedSubs/OutfitCard.jsx
@@ -1,13 +1,0 @@
-
-
-function OutfitCard() {
-
-  return (
-    <div>
-      <div>Outfit Card</div>
-      <div></div>
-    </div>
-  )
-}
-
-export default OutfitCard;
