+++ projectAtelier/client/src/components/RelatedProducts/RelatedSubs/RelatedList.jsx
@@ -1,21 +1,24 @@
 import axios from 'axios';
 import axiosConfig from '../../../Middleware/axiosConfig.js';
-import averageRating from '../../../Middleware/averageRating.js';
 import Box from '@mui/material/Box';
-import {useEffect, useState} from 'react';
 import Carousel from "react-multi-carousel";
 import RelatedCard from './RelatedCard.jsx';
 import "react-multi-carousel/lib/styles.css";
+import AddOutfit from './AddOutfit.jsx';
 import '../../../index.css';
 
 const responsive = {
-  superLargeDesktop: {
-    // the naming can be any, depends on you.
-    breakpoint: { max: 4000, min: 3000 },
+  evenBiggerDesktop: {
+    breakpoint: {max: 2500, min: 2000},
     items: 5
   },
+  superLargeDesktop: {
+    // the naming can be any, depends on you.
+    breakpoint: { max: 2000, min: 1500 },
+    items: 4
+  },
   desktop: {
-    breakpoint: { max: 3000, min: 1024 },
+    breakpoint: { max: 1500, min: 1024 },
     items: 3
   },
   tablet: {
@@ -31,69 +34,49 @@ const responsive = {
 
 
 
-function RelatedList({currentProduct, setProduct}) {
-  const [relatedIDs, setRelatedIDs] = useState([]);
-  const [relatedProducts, setRelatedProducts] = useState([]);
+function RelatedList({currentProduct, setProduct, products, list}) {
 
-  //get all related id's to current product
-  useEffect(() => {
-    if (!!currentProduct.id) {
-      axios.get(axiosConfig.url + '/products/' + currentProduct.id + '/related', axiosConfig).then((response) => {
-        setRelatedIDs(response.data);
-        setRelatedProducts([]);
-        }).catch((err) => {
-          console.log('GET RELATED ID ERROR ', err);
-        })
-
-    }
-  },[currentProduct])
-
-//get all relevant info about each related id
-  useEffect(() => {
-    relatedIDs.forEach((product) => {
-      let options = axiosConfig;
-      options.params ={};
-      options.params.product_id = product;
-      axios.get(options.url + '/reviews/meta', options).then((response) => {
-        let average = averageRating(response.data.ratings);
-        return Number(average);
-      }).then((average) => {
-        axios.get(axiosConfig.url + '/products/' + product, axiosConfig).then((res) => {
-          res.data.average = average;
-          setRelatedProducts(relatedProducts => [...relatedProducts, res.data]);
-        }).catch((err) => {
-          console.log('GET RELATED PRODUCTS ERROR ', err);
-        })
-      })
-    })
-  }, [relatedIDs])
 
   //set the current product when a related card is clicked
-  function handleClick(id) {
+  function relatedClick(id) {
     axios.get(axiosConfig.url + '/products/' + id, axiosConfig).then((response) => {
       setProduct(response.data);
     })
   }
 
-//unique-ify and group related products
-  function removeDupes(array) {
-    let uniques = [...new Map(array.map((product) => [product.id, product])).values()];
-    return uniques;
+  function outfitClick(id) {
+    console.log('I was clicked ', id);
   }
-  let uniqueProds = removeDupes(relatedProducts);
 
-  //sort related products
-  uniqueProds = uniqueProds.sort((a, b) => {
-    return a.id - b.id;
-  })
 
-  return (
+
+  if (list === 'related') {
+    return (
     <div>
-      <div>Related List</div>
+      <div className='related-outfit-header'>Related Products</div>
         <Box className='carousel-box'>
           <Carousel itemClass='carousel-item' responsive={responsive}>
-              {uniqueProds.map((relatedProduct) => {
-                  return <RelatedCard key={relatedProduct.id} product={relatedProduct} currentProduct={currentProduct} handleClick={handleClick} />
+              {products.map((product) => {
+                  return <RelatedCard key={product.id} product={product} currentProduct={currentProduct} handleClick={relatedClick} list={list} />
+              })}
+          </Carousel>
+        </Box>
+
+    </div>
+    );
+  } else {
+    return (
+      <div>
+      <div className='related-outfit-header'>Outfit</div>
+        <Box className='carousel-box'>
+          <Carousel itemClass='carousel-item' responsive={responsive}>
+              {products.map((product, index) => {
+                  if (!product) {
+                    return <AddOutfit key={index}/>
+                  } else {
+                    return <RelatedCard key={product.id} product={product} currentProduct={currentProduct} handleClick={outfitClick} list={list}/>
+                  }
+
               })}
           </Carousel>
         </Box>
@@ -101,5 +84,6 @@ function RelatedList({currentProduct, setProduct}) {
     </div>
     );
   }
+}
 
 export default RelatedList;
